Use python-lmdb package from Ubuntu instead of building our own
Ubuntu 16.04 and 18.04 both have version 0.86, which is the same version
that our py-lmdb submodule was at.

We don't need to support Ubuntu 14.04 anymore (it reaches end of life
next month), although python-lmdb *is* available from trusty-backports
if you need it.

I'm removing this because building it for Python3 fails in CI. I could
fix that but there's no need because we can get it from Ubuntu.